[fix] removed debug statements and added Splashscreen
