+++ app/(auth)/login/withEmail.tsx
@@ -3,7 +3,7 @@ import { useTheme } from "../../../lib/theme/context";
 import { Text, View } from "react-native";
 import Input from "@/components/input";
 import { BtnText, Button, Loader } from "@/components/button";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { router } from "expo-router";
 
 type FormData = {
@@ -12,10 +12,6 @@ type FormData = {
 };
 
 export default function WithEmail() {
-    const renderCount = useRef(0);
-    renderCount.current++;
-    console.log(`withEmail rendered: ${renderCount.current} times`);
-
     useEffect(() => {
         router.prefetch("/(protected)/(tabs)");
     }, []);
