+++ app/(auth)/register/addProfileData.tsx
@@ -5,14 +5,10 @@ import Input from "@/components/input";
 import PhoneNumberInput from "@/components/PhoneNumberInput";
 import { BtnText, Button } from "@/components/button";
 import DateInput from "@/components/dateInput";
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import { router } from "expo-router";
 
 export default function AddProfileDataPage() {
-    const renderCount = useRef(0);
-    renderCount.current++;
-    console.log(`addProfileData rendered: ${renderCount.current} times`);
-
     const { gs, theme } = useTheme();
 
     const [dateSelect, setDateSelect] = useState(false);
