+++ app/(auth)/register/addProfilePicture.tsx
@@ -4,13 +4,9 @@ import { Pressable, Text, View } from "react-native";
 import { Image } from "expo-image";
 import { BtnText, Button } from "@/components/button";
 import { router } from "expo-router";
-import { useEffect, useRef } from "react";
+import { useEffect } from "react";
 
 export default function AddProfilePicturePage() {
-    const renderCount = useRef(0);
-    renderCount.current++;
-    console.log(`addProfilePicture rendered: ${renderCount.current} times`);
-
     const { gs, theme } = useTheme();
 
     useEffect(() => {
