--- hooks/index.d.ts
@@ -1,9 +1,0 @@
-export declare type ErrResp = {
-    message: string;
-    error: any;
-};
-
-export declare type SucResp = {
-    message: string;
-    data: any;
-};
