+++ hooks/useFetch.ts
@@ -1,49 +1,5 @@
-import AsyncStorage from "@react-native-async-storage/async-storage";
-import { useState, useEffect, useCallback } from "react";
-import { ErrResp, SucResp } from "@/hooks/index.d";
-
-const ramStore = new Map<string, any>();
-
-export function useStore<S>(
-    keyValue: string,
-    initialValue?: S,
-): readonly [S, (newVal: S) => void, () => void, boolean] {
-    const key = `store_${keyValue}`;
-
-    const [value, setValue] = useState<S>(() =>
-        ramStore.has(key) ? ramStore.get(key) : initialValue,
-    );
-    const [loaded, setLoaded] = useState(ramStore.has(key));
-
-    useEffect(() => {
-        if (loaded) return;
-        AsyncStorage.getItem(key).then((stored) => {
-            if (stored !== null) {
-                const parsed = JSON.parse(stored) as S;
-                ramStore.set(key, parsed);
-                setValue(parsed);
-            }
-            setLoaded(true);
-        });
-    }, []);
-
-    const setFunc = useCallback(
-        (newVal: S) => {
-            ramStore.set(key, newVal);
-            setValue(newVal);
-            AsyncStorage.setItem(key, JSON.stringify(newVal)).catch(console.error);
-        },
-        [key],
-    );
-
-    const clearFunc = useCallback(() => {
-        ramStore.delete(key);
-        setValue(initialValue!);
-        AsyncStorage.removeItem(key).catch(console.error);
-    }, [key]);
-
-    return [value, setFunc, clearFunc, loaded] as const;
-}
+import { useStore } from "@/hooks/useStore";
+import { useEffect, useState } from "react";
 
 export function useFetch<S>(route: string, obj: RequestInit) {
     const [fetchData, setFetchData, free] = useStore<{
@@ -78,13 +34,13 @@ export function useFetch<S>(route: string, obj: RequestInit) {
                 const req = await fetch(route, obj);
 
                 if (!req.ok) {
-                    const err = (await req.json()) as ErrResp;
+                    const err = await req.json();
                     setError(new Error(err.message));
                     setLoading(false);
                     return;
                 }
 
-                const data = (await req.json()) as SucResp;
+                const data = await req.json();
 
                 setFetchData({
                     data: data as S,
